# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "InvoicePro API Documentation",
    "version": "1.0.0",
    "description": "Exhaustive OpenAPI 3.0 specification for all InvoicePro backend API endpoints (Authentication, Customer Management, Inventory & Stock adjustments, Invoice lifecycle workflows, and Admin User management)."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/customers": {
      "get": {
        "tags": ["Customers"],
        "summary": "List Customers",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Search by customer name"
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers"
          }
        }
      },
      "post": {
        "tags": ["Customers"],
        "summary": "Create Customer",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Enterprise Corp" },
                  "email": {
                    "type": "string",
                    "example": "billing@enterprise.com"
                  },
                  "phone": { "type": "string" },
                  "billing_address": { "type": "string" },
                  "gst_number": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created successfully"
          }
        }
      }
    },
...
    "/api/inventory": {
      "get": {
        "tags": ["Inventory"],
        "summary": "List Inventory Products",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          {
            "name": "is_active",
            "in": "query",
            "schema": { "type": "boolean" }
          },
          {
            "name": "low_stock",
            "in": "query",
            "schema": { "type": "boolean" }
          },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "List of products"
          }
        }
      },
      "post": {
        "tags": ["Inventory"],
        "summary": "Create Product",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Premium Product" },
                  "description": { "type": "string" },
                  "sku": { "type": "string", "example": "PREM-001" },
                  "unit": { "type": "string", "example": "pcs" },
                  "category": { "type": "string" },
                  "price": { "type": "number", "example": 150.0 },
                  "cost": { "type": "number" },
                  "tax_rate": { "type": "number" },
                  "stock": { "type": "integer" },
                  "reorder_level": { "type": "integer" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully"
          }
        }
      }
    },
...
    "/api/invoices": {
      "get": {
        "tags": ["Invoices"],
        "summary": "List Invoices",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "customer", "in": "query", "schema": { "type": "string" } },
          { "name": "from", "in": "query", "schema": { "type": "string" } },
          { "name": "to", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of invoices" }
        }
      },
      "post": {
        "tags": ["Invoices"],
        "summary": "Create Invoice (Draft)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "invoice_number": { "type": "string" },
                  "customer_id": { "type": "string" },
                  "customer_name": { "type": "string" },
                  "customer_email": { "type": "string" },
                  "invoice_date": { "type": "string" },
                  "due_date": { "type": "string" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "product_id": { "type": "string" },
                        "quantity": { "type": "number" },
                        "price": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Invoice created as Draft" }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List Users",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of system users" }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create User (Admin-Only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "role_id": { "type": "string" }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" }
        }
      }
    },
    "/api/users/{id}": {
      "put": {
        "tags": ["Users"],
        "summary": "Update User",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Deactivate User",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "User deactivated successfully" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}